Refactoring: GeoDataRiverSurvey::TranslateRiverPathPointCommand is moved to separated files.
