+++ libs/geodata/riversurvey/geodatariversurvey.cpp
@@ -18,6 +18,7 @@
 #include "private/geodatariversurvey_mouseshiftriverpathcentercommand.h"
 #include "private/geodatariversurvey_removeextensioncommand.h"
 #include "private/geodatariversurvey_setdisplaysettingcommand.h"
+#include "private/geodatariversurvey_translateriverpathpointcommand.h"
 
 #include <guicore/pre/base/preprocessorwindowinterface.h>
 #include <guicore/base/iricmainwindowinterface.h>
@@ -638,74 +639,6 @@ void GeoDataRiverSurvey::mouseDoubleClickEvent(QMouseEvent* /*event*/, PreProces
 	}
 	openCrossSectionWindow();
 }
-
-class GeoDataRiverSurvey::TranslateRiverPathPointCommand : public QUndoCommand
-{
-public:
-	TranslateRiverPathPointCommand(QPoint from, QPoint to, GeoDataRiverSurvey* data) :
-		QUndoCommand {GeoDataRiverSurvey::tr("Move Traversal Line")}
-	{
-		PreProcessorGraphicsViewInterface* gview = data->graphicsView();
-		m_rs = data;
-		double fromX, fromY, toX, toY;
-		fromX = from.x();
-		fromY = from.y();
-		toX = to.x();
-		toY = to.y();
-		gview->viewportToWorld(fromX, fromY);
-		gview->viewportToWorld(toX, toY);
-		QVector2D offset(toX - fromX, toY - fromY);
-
-		GeoDataRiverPathPoint* p = data->m_headPoint->nextPoint();
-		while (p != nullptr) {
-			if (p->IsSelected) {
-				m_points.append(p);
-				m_oldPositions.append(p->position());
-				m_newPositions.append(p->position() + offset);
-			}
-			p = p->nextPoint();
-		}
-	}
-	void redo() {
-		m_rs->m_gridThread->cancel();
-		for (int i = 0; i < m_points.count(); ++i) {
-			m_points[i]->setPosition(m_newPositions[i]);
-		}
-		m_rs->headPoint()->updateRiverShapeInterpolators();
-		m_rs->updateShapeData();
-		m_rs->renderGraphicsView();
-		m_rs->updateCrossectionWindows();
-		m_rs->setMapped(false);
-	}
-	void undo() {
-		m_rs->m_gridThread->cancel();
-		for (int i = 0; i < m_points.count(); ++i) {
-			m_points[i]->setPosition(m_oldPositions[i]);
-		}
-		m_rs->headPoint()->updateRiverShapeInterpolators();
-		m_rs->updateShapeData();
-		m_rs->renderGraphicsView();
-		m_rs->updateCrossectionWindows();
-	}
-	int id() const {
-		return iRIC::generateCommandId("GeoDataRiverSurveyTranslateCommand");
-	}
-	virtual bool mergeWith(const QUndoCommand* other) {
-		const TranslateRiverPathPointCommand* other2 = dynamic_cast<const TranslateRiverPathPointCommand*>(other);
-		if (other2 == nullptr) { return false; }
-		if (other2->m_points == m_points) {
-			m_newPositions = other2->m_newPositions;
-			return true;
-		}
-		return false;
-	}
-
-private:
-	QList<GeoDataRiverPathPoint*> m_points;
-	QList<QVector2D> m_oldPositions;
-	QList<QVector2D> m_newPositions;
-	GeoDataRiverSurvey* m_rs;
-};
 
 class GeoDataRiverSurvey::MouseMoveExtensionCommand : public QUndoCommand
 {
